pyproject.toml
[project]
name = "multi-agent-customer-service"
version = "0.1.0"
description = "基于 Orchestrator-Worker 架构的多智能体客服系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.1,<4.1",
    "celery>=5.6.2",
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.5",
    "httpx>=0.28.1",
    "langchain>=1.2.9",
    "langchain-anthropic>=1.3.2",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "neo4j>=6.1.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pymilvus>=2.6.8",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis[hiredis]>=6.4.0",
    "sqlalchemy>=2.0.46",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.40.0",
    "websockets>=16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.28",
    "ruff>=0.5",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
